Use cli config.yml + handle multiple gateways and auth
This commit dumps the openfaas-bitbar specific config.yaml file in
.openfaas and instead uses the config file that was later introduced to
the CLI, config.yml.

This enables the openfaas-bitbar app to use the basic auth config from
`faas login` when connecting to a gateway endpoint.

It also allows the plugin to talk to and display data from multiple
gateways.

The data returned from each function and the rendered info has been
updated.

One negative impact of the switch to the CLI config is that we now lose
the Prometheus endpoint configuration, so graphs have been removed. This
may be revisted in a future update (likely requiring updates to the
CLI).